Gabor Toth:
A repaint in FedEx livery is ready and avaliable at:

  http://fgfs.i-net.hu/downloads/MD11.tar.bz2

I've copied the KLM related files and repainted the textures only. The 3D
model also should be changed, because it's colour is gray, but I can't do
that.

The paintwork is my own work using The Gimp. Only FedEx logo was downloaded
from the Web. I think I may publish the textures under GPL.

An updated version will follow if I found the way to make the cowling around
engine #2 look like the real livery.